Move `background_archiving()` and surrounding functions/data structures into own `archiving` module without any logic
