+++ crates/subspace-farmer/src/plotting.rs
@@ -1,26 +1,19 @@
 #[cfg(test)]
 mod tests;
 
+use crate::archiving::PiecesToPlot;
 use crate::commitments::Commitments;
 use crate::object_mappings::ObjectMappings;
 use crate::plot::Plot;
 use crate::rpc::RpcClient;
-use futures::channel::mpsc;
-use futures::{SinkExt, StreamExt};
-use log::{debug, error, info, warn};
-use std::sync::atomic::{AtomicU32, Ordering};
+use log::error;
 use std::sync::Arc;
 use std::time::Duration;
-use subspace_archiving::archiver::{ArchivedSegment, Archiver};
-use subspace_core_primitives::objects::{GlobalObject, PieceObject, PieceObjectMapping};
-use subspace_core_primitives::{FlatPieces, PieceIndex, Sha256Hash};
-use subspace_rpc_primitives::{EncodedBlockWithObjectMapping, FarmerMetadata};
+use subspace_core_primitives::{FlatPieces, PieceIndex};
+use subspace_rpc_primitives::FarmerMetadata;
 use subspace_solving::{BatchEncodeError, SubspaceCodec};
 use thiserror::Error;
-use tokio::sync::oneshot::Receiver;
 use tokio::{sync::oneshot, task::JoinHandle};
-
-const BEST_BLOCK_REQUEST_TIMEOUT: Duration = Duration::from_secs(60);
 
 #[derive(Debug, Error)]
 pub enum PlottingError {
@@ -36,12 +29,6 @@ pub enum PlottingError {
     JoinTask(tokio::task::JoinError),
     #[error("Archiver instantiation error: {0}")]
     Archiver(subspace_archiving::archiver::ArchiverInstantiationError),
-}
-
-struct PiecesToPlot {
-    /// Offset of the index of the first piece in `pieces`
-    piece_index_offset: u64,
-    pieces: FlatPieces,
 }
 
 /// `Plotting` struct is the abstraction of the plotting process
@@ -101,7 +88,7 @@ impl Plotting {
             std::sync::mpsc::sync_channel::<PiecesToPlot>(0);
 
         // Get a handle for the background task, so that we can wait on it later if we want to
-        let archiving_handle = tokio::spawn(background_archiving(
+        let archiving_handle = tokio::spawn(crate::archiving::background_archiving(
             pieces_to_plot_sender,
             plot,
             metadata,
@@ -158,291 +145,6 @@ impl Drop for Plotting {
     }
 }
 
-// TODO: Blocks that are coming form substrate node are fully trusted right now, which we probably
-//  don't want eventually
-/// Maintains plot in up to date state plotting new pieces as they are produced on the network.
-async fn background_archiving<T: RpcClient + Clone + Send + 'static>(
-    pieces_to_plot_sender: std::sync::mpsc::SyncSender<PiecesToPlot>,
-    plot: Plot,
-    metadata: FarmerMetadata,
-    object_mappings: ObjectMappings,
-    client: T,
-    best_block_number_check_interval: Duration,
-    mut stop_receiver: Receiver<()>,
-) -> Result<(), PlottingError> {
-    let weak_plot = plot.downgrade();
-    let FarmerMetadata {
-        confirmation_depth_k,
-        record_size,
-        recorded_history_segment_size,
-        ..
-    } = metadata;
-
-    // TODO: This assumes fixed size segments, which might not be the case
-    let merkle_num_leaves = u64::from(recorded_history_segment_size / record_size * 2);
-
-    let maybe_last_root_block = tokio::task::spawn_blocking({
-        let plot = plot.clone();
-
-        move || plot.get_last_root_block().map_err(PlottingError::LastBlock)
-    })
-    .await
-    .unwrap()?;
-
-    let mut archiver = if let Some(last_root_block) = maybe_last_root_block {
-        // Continuing from existing initial state
-        if plot.is_empty() {
-            return Err(PlottingError::ContinueError);
-        }
-
-        let last_archived_block_number = last_root_block.last_archived_block().number;
-        info!("Last archived block {}", last_archived_block_number);
-
-        let maybe_last_archived_block = client
-            .block_by_number(last_archived_block_number)
-            .await
-            .map_err(PlottingError::RpcError)?;
-
-        match maybe_last_archived_block {
-            Some(EncodedBlockWithObjectMapping {
-                block,
-                object_mapping,
-            }) => Archiver::with_initial_state(
-                record_size as usize,
-                recorded_history_segment_size as usize,
-                last_root_block,
-                &block,
-                object_mapping,
-            )
-            .map_err(PlottingError::Archiver)?,
-            None => {
-                return Err(PlottingError::GetBlockError(last_archived_block_number));
-            }
-        }
-    } else {
-        // Starting from genesis
-        if !plot.is_empty() {
-            // Restart before first block was archived, erase the plot
-            // TODO: Erase plot
-        }
-
-        Archiver::new(record_size as usize, recorded_history_segment_size as usize)
-            .map_err(PlottingError::Archiver)?
-    };
-
-    drop(plot);
-
-    let (new_block_to_archive_sender, new_block_to_archive_receiver) =
-        std::sync::mpsc::sync_channel::<Arc<AtomicU32>>(0);
-
-    // Process blocks since last fully archived block (or genesis) up to the current head minus K
-    let mut blocks_to_archive_from = archiver
-        .last_archived_block_number()
-        .map(|n| n + 1)
-        .unwrap_or_default();
-
-    // Erasure coding in archiver and piece encoding are CPU-intensive operations.
-    tokio::task::spawn_blocking({
-        let client = client.clone();
-        let weak_plot = weak_plot.clone();
-
-        #[allow(clippy::mut_range_bound)]
-        move || {
-            let runtime_handle = tokio::runtime::Handle::current();
-            info!("Plotting new blocks in the background");
-
-            'outer: for blocks_to_archive_to in new_block_to_archive_receiver.into_iter() {
-                let blocks_to_archive_to = blocks_to_archive_to.load(Ordering::Relaxed);
-                if blocks_to_archive_to >= blocks_to_archive_from {
-                    debug!(
-                        "Archiving blocks {}..={}",
-                        blocks_to_archive_from, blocks_to_archive_to,
-                    );
-                }
-
-                for block_to_archive in blocks_to_archive_from..=blocks_to_archive_to {
-                    let EncodedBlockWithObjectMapping {
-                        block,
-                        object_mapping,
-                    } = match runtime_handle.block_on(client.block_by_number(block_to_archive)) {
-                        Ok(Some(block)) => block,
-                        Ok(None) => {
-                            error!(
-                                "Failed to get block #{} from RPC: Block not found",
-                                block_to_archive,
-                            );
-
-                            blocks_to_archive_from = block_to_archive;
-                            continue 'outer;
-                        }
-                        Err(error) => {
-                            error!(
-                                "Failed to get block #{} from RPC: {}",
-                                block_to_archive, error,
-                            );
-
-                            blocks_to_archive_from = block_to_archive;
-                            continue 'outer;
-                        }
-                    };
-
-                    let mut last_root_block = None;
-                    for archived_segment in archiver.add_block(block, object_mapping) {
-                        let ArchivedSegment {
-                            root_block,
-                            pieces,
-                            object_mapping,
-                        } = archived_segment;
-                        let piece_index_offset = merkle_num_leaves * root_block.segment_index();
-
-                        let pieces_to_plot = PiecesToPlot {
-                            piece_index_offset,
-                            pieces,
-                        };
-                        if let Err(_error) = pieces_to_plot_sender.send(pieces_to_plot) {
-                            // No receivers, exiting
-                            break 'outer;
-                        }
-
-                        let object_mapping =
-                            create_global_object_mapping(piece_index_offset, object_mapping);
-
-                        if let Err(error) = object_mappings.store(&object_mapping) {
-                            error!("Failed to store object mappings for pieces: {}", error);
-                        }
-                        let segment_index = root_block.segment_index();
-                        last_root_block.replace(root_block);
-
-                        info!(
-                            "Archived segment {} at block {}",
-                            segment_index, block_to_archive
-                        );
-                    }
-
-                    if let Some(last_root_block) = last_root_block {
-                        if let Some(plot) = weak_plot.upgrade() {
-                            if let Err(error) = plot.set_last_root_block(&last_root_block) {
-                                error!("Failed to store last root block: {}", error);
-                            }
-                        }
-                    }
-                }
-
-                blocks_to_archive_from = blocks_to_archive_to + 1;
-            }
-        }
-    });
-
-    info!("Subscribing to new heads");
-    let mut new_head = client
-        .subscribe_new_head()
-        .await
-        .map_err(PlottingError::RpcError)?;
-
-    let block_to_archive = Arc::new(AtomicU32::default());
-
-    if maybe_last_root_block.is_none() {
-        // If not continuation, archive genesis block
-        new_block_to_archive_sender
-            .send(Arc::clone(&block_to_archive))
-            .expect("Failed to send genesis block archiving message");
-    }
-
-    let (mut best_block_number_sender, mut best_block_number_receiver) = mpsc::channel(1);
-
-    tokio::spawn(async move {
-        loop {
-            tokio::time::sleep(best_block_number_check_interval).await;
-
-            // In case connection dies, we need to disconnect from the node
-            let best_block_number_result =
-                tokio::time::timeout(BEST_BLOCK_REQUEST_TIMEOUT, client.best_block_number()).await;
-
-            let is_error = !matches!(best_block_number_result, Ok(Ok(_)));
-            // Result doesn't matter here
-            let _ = best_block_number_sender
-                .send(best_block_number_result)
-                .await;
-
-            if is_error {
-                break;
-            }
-        }
-    });
-
-    let mut last_best_block_number_error = false;
-
-    // Listen for new blocks produced on the network
-    loop {
-        tokio::select! {
-            _ = &mut stop_receiver => {
-                info!("Plotting stopped!");
-                break;
-            }
-            result = new_head.recv() => {
-                match result {
-                    Some(head) => {
-                        let block_number = u32::from_str_radix(&head.number[2..], 16).unwrap();
-                        debug!("Last block number: {:#?}", block_number);
-
-                        if let Some(block_number) = block_number.checked_sub(confirmation_depth_k) {
-                            // We send block that should be archived over channel that doesn't have
-                            // a buffer, atomic integer is used to make sure archiving process
-                            // always read up to date value
-                            block_to_archive.store(block_number, Ordering::Relaxed);
-                            let _ = new_block_to_archive_sender.try_send(Arc::clone(&block_to_archive));
-                        }
-                    },
-                    None => {
-                        debug!("Subscription has forcefully closed from node side!");
-                        break;
-                    }
-                }
-            }
-            maybe_result = best_block_number_receiver.next() => {
-                match maybe_result {
-                    Some(Ok(Ok(best_block_number))) => {
-                        debug!("Best block number: {:#?}", best_block_number);
-                        last_best_block_number_error = false;
-
-                        if let Some(block_number) = best_block_number.checked_sub(confirmation_depth_k) {
-                            // We send block that should be archived over channel that doesn't have
-                            // a buffer, atomic integer is used to make sure archiving process
-                            // always read up to date value
-                            block_to_archive.fetch_max(block_number, Ordering::Relaxed);
-                            let _ = new_block_to_archive_sender.try_send(Arc::clone(&block_to_archive));
-                        }
-                    }
-                    Some(Ok(Err(error))) => {
-                        if last_best_block_number_error {
-                            error!("Request to get new best block failed second time: {error}");
-                            break;
-                        } else {
-                            warn!("Request to get new best block failed: {error}");
-                            last_best_block_number_error = true;
-                        }
-                    }
-                    Some(Err(_error)) => {
-                        if last_best_block_number_error {
-                            error!("Request to get new best block timed out second time");
-                            break;
-                        } else {
-                            warn!("Request to get new best block timed out");
-                            last_best_block_number_error = true;
-                        }
-                    }
-                    None => {
-                        debug!("Best block number channel closed!");
-                        break;
-                    }
-                }
-            }
-        }
-    }
-
-    Ok(())
-}
-
 /// Plot a set of pieces into a particular plot and commitment database.
 fn plot_pieces(
     subspace_codec: &mut SubspaceCodec,
@@ -475,26 +177,4 @@ fn plot_pieces(
     }
 
     Ok(())
-}
-
-fn create_global_object_mapping(
-    piece_index_offset: u64,
-    object_mapping: Vec<PieceObjectMapping>,
-) -> Vec<(Sha256Hash, GlobalObject)> {
-    object_mapping
-        .iter()
-        .enumerate()
-        .flat_map(move |(position, object_mapping)| {
-            object_mapping.objects.iter().map(move |piece_object| {
-                let PieceObject::V0 { hash, offset } = piece_object;
-                (
-                    *hash,
-                    GlobalObject::V0 {
-                        piece_index: piece_index_offset + position as u64,
-                        offset: *offset,
-                    },
-                )
-            })
-        })
-        .collect()
 }
